requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
configobj==5.0.6
selenium==3.141.0
pluggy==0.13.1
pytest==6.1.2
pytest-assume==2.4.2
pytest-bdd==4.0.1
pytest-check==0.3.9
pytest-logging==2015.11.4
pytest-runner==5.2
pytest-steps==1.7.3
pytest-subtests==0.4.0
pytest-testrail==2.9.0
paramiko==2.7.2
pymongo==3.8.0


requests~=2.22.0